+++ packages/core/test/support/factory.js
@@ -14,7 +14,7 @@ export function page (id, attrs = {}) {
   attrs.path = attrs.path || (id === 'index' ? '' : id)
   attrs.title = attrs.title || id
   attrs.content = attrs.content || ''
-  attrs.type = attrs.type || (attrs.template ? 'template' : (attrs.tokens ? 'tokens' : 'documentation'))
+  attrs.type = attrs.type || (attrs.tokens ? 'tokens' : (attrs.template ? 'template' : 'documentation'))
   attrs.files = attrs.files || []
   attrs.childIds = attrs.childIds || []
   attrs.componentIds = attrs.componentIds || []